+++ src/query/validation.c
@@ -29,7 +29,7 @@
   } while (0)
 
 static void verify_where(Query *query);
-static void verify_rtable(Query *query);
+static void verify_select_targets(Query *query);
 static void verify_aggregators(Query *query);
 static void verify_non_system_column(Var *var);
 static bool option_matches(DefElem *option, char *name, bool value);
@@ -106,24 +106,64 @@ void verify_anonymization_requirements(Query *query)
 
   verify_where(query);
   verify_aggregators(query);
-  verify_rtable(query);
+  verify_select_targets(query);
 }
 
-static void verify_rtable(Query *query)
+static void verify_select_targets(Query *query)
 {
-  NOT_SUPPORTED(list_length(query->rtable) > 1, "JOINs in anonymizing queries");
-
   ListCell *cell = NULL;
+
   foreach (cell, query->rtable)
   {
     RangeTblEntry *range_table = lfirst_node(RangeTblEntry, cell);
-    NOT_SUPPORTED(range_table->rtekind == RTE_SUBQUERY, "Subqueries in anonymizing queries");
-    NOT_SUPPORTED(range_table->rtekind == RTE_JOIN, "JOINs in anonymizing queries");
 
-    if (range_table->rtekind == RTE_RELATION)
-      NOT_SUPPORTED(has_subclass(range_table->relid) || has_superclass(range_table->relid), "Inheritance in anonymizing queries.");
-    else
-      FAILWITH("Unsupported FROM clause.");
+    switch (range_table->rtekind)
+    {
+    case RTE_RELATION:
+      NOT_SUPPORTED(has_subclass(range_table->relid) || has_superclass(range_table->relid), "Inheritance in anonymizing queries");
+      break;
+
+    case RTE_JOIN:
+      NOT_SUPPORTED(range_table->jointype == JOIN_SEMI || range_table->jointype == JOIN_ANTI, "SEMI JOIN in anonymizing queries");
+      break;
+
+    case RTE_SUBQUERY:
+      FAILWITH_CODE(ERRCODE_FEATURE_NOT_SUPPORTED, "Subqueries in anonymizing queries are not supported.");
+      break;
+
+    default:
+      FAILWITH_CODE(ERRCODE_FEATURE_NOT_SUPPORTED, "Unsupported FROM clause.");
+      break;
+    }
+  }
+
+  foreach (cell, query->jointree->fromlist)
+  {
+    if (list_length(query->jointree->fromlist) == 1 && IsA(lfirst(cell), RangeTblRef))
+      break;
+
+    NOT_SUPPORTED(IsA(lfirst(cell), RangeTblRef) || IsA(lfirst(cell), FromExpr), "CROSS JOIN in anonymizing queries");
+
+    Assert(IsA(lfirst(cell), JoinExpr));
+    JoinExpr *join_expr = lfirst_node(JoinExpr, cell);
+
+    List *subjects = NIL, *targets = NIL;
+    collect_equalities_from_filters(join_expr->quals, &subjects, &targets);
+
+    ListCell *subject_cell = NULL, *target_cell = NULL;
+    forboth(subject_cell, subjects, target_cell, targets)
+    {
+      Node *subject_expression = unwrap_cast(lfirst(subject_cell));
+      Node *target_expression = unwrap_cast(lfirst(target_cell));
+
+      if (!IsA(subject_expression, Var))
+        FAILWITH_CODE(ERRCODE_FEATURE_NOT_SUPPORTED, "Left side of equality in pre-anonymization JOIN filter has to be a simple column reference.");
+      if (!IsA(target_expression, Var))
+        FAILWITH_CODE(ERRCODE_FEATURE_NOT_SUPPORTED, "Right side of equality in pre-anonymization JOIN filter has to be a simple column reference.");
+    }
+
+    list_free(subjects);
+    list_free(targets);
   }
 }
 
@@ -188,6 +228,11 @@ static bool verify_aggregator(Node *node, void *context)
         aggoid != g_oid_cache.is_suppress_bin)
       FAILWITH_LOCATION(aggref->location, "Unsupported aggregate in query.");
 
+    if ((aggoid == g_oid_cache.sum_noise ||
+         aggoid == g_oid_cache.avg_noise) &&
+        TypeCategory(linitial_oid(aggref->aggargtypes)) == TYPCATEGORY_DATETIME)
+      FAILWITH_LOCATION(aggref->location, "Unsupported aggregate in query.");
+
     if (aggoid == g_oid_cache.count_value || aggoid == g_oid_cache.count_value_noise ||
         is_sum_oid(aggoid) || aggoid == g_oid_cache.sum_noise ||
         is_avg_oid(aggoid) || aggoid == g_oid_cache.avg_noise)
@@ -226,10 +271,11 @@ static void verify_bucket_expression(Node *node)
   if (IsA(node, FuncExpr))
   {
     FuncExpr *func_expr = (FuncExpr *)node;
-    if (is_allowed_cast(func_expr->funcid))
+    if (is_allowed_cast(func_expr))
     {
       Assert(list_length(func_expr->args) == 1); /* All allowed casts require exactly one argument. */
       verify_bucket_expression(linitial(func_expr->args));
+      return;
     }
 
     if (!is_allowed_function(func_expr->funcid))
@@ -237,11 +283,14 @@ static void verify_bucket_expression(Node *node)
 
     Assert(list_length(func_expr->args) > 0); /* All allowed functions require at least one argument. */
 
-    if (!IsA(unwrap_cast(linitial(func_expr->args)), Var))
+    int primary_arg = primary_arg_index(func_expr->funcid);
+    if (!IsA(unwrap_cast(list_nth(func_expr->args, primary_arg)), Var))
       FAILWITH_LOCATION(func_expr->location, "Primary argument for a generalization function has to be a simple column reference.");
 
-    for (int i = 1; i < list_length(func_expr->args); i++)
+    for (int i = 0; i < list_length(func_expr->args); i++)
     {
+      if (i == primary_arg)
+        continue;
       Node *arg = unwrap_cast((Node *)list_nth(func_expr->args, i));
       if (!is_stable_expression(arg))
         FAILWITH_LOCATION(exprLocation(arg), "Non-primary arguments for a generalization function have to be simple constants.");
@@ -376,7 +425,7 @@ double numeric_value_to_double(Oid type, Datum value)
 
 static bool option_matches(DefElem *option, char *name, bool value)
 {
-  return strcasecmp(option->defname, name) == 0 && defGetBoolean(option) == value;
+  return pg_strcasecmp(option->defname, name) == 0 && defGetBoolean(option) == value;
 }
 
 static bool is_equality_op(Oid opno)
@@ -414,7 +463,7 @@ void collect_equalities_from_filters(Node *node, List **subjects, List **targets
     }
   }
 
-  FAILWITH("Only equalities between generalization expressions and constants are allowed as pre-anonymization filters.");
+  FAILWITH("Only equalities are allowed in pre-anonymization filters.");
 }
 
 static Var *get_bucket_expression_column_ref(Node *bucket_expression)
@@ -424,7 +473,9 @@ static Var *get_bucket_expression_column_ref(Node *bucket_expression)
     return (Var *)bucket_expression;
   /* If the bucket expression is not a direct column reference, it means it is a simple function call. */
   FuncExpr *func_expr = castNode(FuncExpr, bucket_expression);
-  return castNode(Var, unwrap_cast(linitial(func_expr->args)));
+
+  int primary_arg = primary_arg_index(func_expr->funcid);
+  return castNode(Var, unwrap_cast(list_nth(func_expr->args, primary_arg)));
 }
 
 static void verify_column_usage_in_filter(AccessLevel access_level, Node *bucket_expression, List *range_tables)