+++ test/sql/misc.sql
@@ -101,6 +101,9 @@ SELECT FROM test_customers WHERE FALSE;
 SELECT TRUE AS "_" FROM test_customers WHERE 1 <> 1 LIMIT 0;
 SELECT id FROM test_customers WHERE NULL = NULL;
 
+-- JOIN between personal tables produces multiple AIDs
+EXPLAIN VERBOSE SELECT COUNT(*) FROM test_customers c JOIN test_purchases pur ON c.id = cid;
+
 -- Tolerate `diffix.agg_noise` in direct access level
 SET pg_diffix.session_access_level = 'direct';
 SELECT diffix.sum_noise(discount), diffix.count_noise(*) FROM test_customers;